refactor(viewer): split StorageTab into components/admin/adminStorage
1,407-line tab becomes a folder: thin StorageTab on AdminTabShell +
DataTable (custom row renderer for folder/source rows), StorageToolbar,
OverridesPanel, SourceRow, SourceCard, FolderRows, conversionOverrides,
useResizableColumns, format helpers, and a useAdminStorage hook owning
the listing, selection, folder state and every DL/Convert/Delete/Move
action. DataTable gains `renderRow` for rows that are not plain cell
lists. Markup, class names and strings unchanged.
